+++ api/src/main/java/com/albert/commerce/common/handler/GlobalExceptionHandler.java
@@ -2,8 +2,8 @@ package com.albert.commerce.common.handler;
 
 import com.albert.commerce.common.exception.ErrorResponse;
 import com.albert.commerce.common.units.BusinessLinks;
+import com.albert.commerce.product.ProductNotFoundException;
 import com.albert.commerce.product.UnauthorizedModificationException;
-import com.albert.commerce.product.infra.persistence.ProductNotFoundException;
 import com.albert.commerce.store.MyStoreNotFoundException;
 import com.albert.commerce.store.StoreNotFoundException;
 import com.albert.commerce.store.command.application.StoreAlreadyExistsException;
