+++ api/src/main/java/com/albert/commerce/user/query/domain/UserDao.java
@@ -6,6 +6,4 @@ import java.util.Optional;
 public interface UserDao {
 
     Optional<User> findByEmail(String email);
-
-    boolean existsByEmail(String email);
 }
