+++ api/src/test/java/com/albert/commerce/comment/ui/CommentControllerTest.java
@@ -17,13 +17,13 @@ import com.albert.commerce.comment.command.application.CommentResponse;
 import com.albert.commerce.comment.command.application.CommentService;
 import com.albert.commerce.comment.query.domain.CommentDao;
 import com.albert.commerce.common.infra.persistence.Money;
+import com.albert.commerce.product.ProductNotFoundException;
 import com.albert.commerce.product.command.application.ProductRequest;
 import com.albert.commerce.product.command.application.dto.ProductCreatedResponse;
 import com.albert.commerce.product.command.application.dto.ProductService;
 import com.albert.commerce.product.command.domain.Product;
 import com.albert.commerce.product.command.domain.ProductId;
-import com.albert.commerce.product.infra.persistence.ProductNotFoundException;
-import com.albert.commerce.product.query.ProductDao;
+import com.albert.commerce.product.query.domain.ProductDao;
 import com.albert.commerce.store.StoreNotFoundException;
 import com.albert.commerce.store.command.application.SellerStoreService;
 import com.albert.commerce.store.command.application.dto.NewStoreRequest;
